Rename packages for `pkg/skuba/actions/node/*`
The packages for the subfolders were all grouped under the `node` package.
Instead they now are part a package with the same name as the last folder they are in.
