+++ pkg/skuba/actions/node/bootstrap/bootstrap.go
@@ -15,7 +15,7 @@
  *
  */
 
-package node
+package bootstrap
 
 import (
 	"fmt"
