+++ pkg/skuba/actions/node/join/config.go
@@ -15,7 +15,7 @@
  *
  */
 
-package node
+package join
 
 import (
 	"io/ioutil"
