+++ pkg/skuba/actions/node/join/config_test.go
@@ -15,7 +15,7 @@
  *
  */
 
-package node
+package join
 
 import (
 	"reflect"
