+++ pkg/skuba/actions/node/join/join.go
@@ -15,7 +15,7 @@
  *
  */
 
-package node
+package join
 
 import (
 	"fmt"
