+++ pkg/skuba/actions/node/remove/remove.go
@@ -15,7 +15,7 @@
  *
  */
 
-package node
+package remove
 
 import (
 	"fmt"
